Nothing to see here :D
Ever get tired of explaining setup() and draw() methods to the fifth
student that day that has just told you that they hate Processing?
This update is for you!  Now there is a Processing Tips page that
covers the fundamentals, with examples.  Currently it is only
accessible from the Googly Eyes recipe.  I wonder how long until
someone notices….